app/api/pyproject.toml
[project]
name = "django-login"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "argon2>=0.1.10",
    "argon2-cffi>=25.1.0",
    "crypto>=1.4.1",
    "cryptography>=45.0.4",
    "django-ninja>=1.4.0",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.2",
    "pyjwt>=2.10.1",
    "ruff>=0.11.4",
]
